+++ src/main/java/plus/dragons/createcentralkitchen/client/CCKClient.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2025  DragonsPlus
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package plus.dragons.createcentralkitchen.client;
+
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import plus.dragons.createcentralkitchen.client.model.CCKPartialModels;
+import plus.dragons.createcentralkitchen.common.CCKCommon;
+import plus.dragons.createcentralkitchen.integration.ModIntegration;
+
+@Mod(value = CCKCommon.ID, dist = Dist.CLIENT)
+public class CCKClient {
+    public CCKClient(IEventBus modBus, ModContainer modContainer) {
+        CCKPartialModels.register();
+        modBus.register(this);
+    }
+
+    @SubscribeEvent
+    public void onClientSetup(final FMLClientSetupEvent event) {
+        for (ModIntegration integration : ModIntegration.values()) {
+            if (integration.enabled())
+                event.enqueueWork(integration::onClientSetup);
+        }
+    }
+}