+++ src/main/java/plus/dragons/createcentralkitchen/mixin/farmersdelight/client/BlazeBurnerBlockEntityMixin.java
@@ -16,10 +16,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package plus.dragons.createcentralkitchen.mixin.farmersdelight;
+package plus.dragons.createcentralkitchen.mixin.farmersdelight.client;
 
-import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import com.llamalad7.mixinextras.sugar.Local;
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.simibubi.create.content.processing.burner.BlazeBurnerBlockEntity;
 import com.simibubi.create.foundation.blockEntity.SmartBlockEntity;
 import me.fallenbreath.conditionalmixin.api.annotation.Condition;
@@ -28,27 +27,43 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import plus.dragons.createcentralkitchen.integration.ModIntegration;
+import plus.dragons.createcentralkitchen.integration.farmersdelight.burner.ChefBlazeBurnerBlockEntity;
 import vectorwing.farmersdelight.common.block.entity.HeatableBlockEntity;
 import vectorwing.farmersdelight.common.tag.ModTags;
 
 @Restriction(require = @Condition(ModIntegration.Constants.FARMERSDELIGHT))
 @Mixin(BlazeBurnerBlockEntity.class)
-public abstract class BlazeBurnerBlockEntityMixin extends SmartBlockEntity {
+public abstract class BlazeBurnerBlockEntityMixin extends SmartBlockEntity implements ChefBlazeBurnerBlockEntity {
+    @Unique
+    private boolean chef;
+
     private BlazeBurnerBlockEntityMixin(BlockEntityType<?> type, BlockPos pos, BlockState state) {
         super(type, pos, state);
     }
 
-    @ModifyReturnValue(method = "isValidBlockAbove", at = @At("TAIL"))
-    private boolean isHeatableBlockEntityAbove(boolean original, @Local BlockState aboveState) {
-        if (original)
-            return true;
+    @Override
+    public boolean isChef() {
+        return chef;
+    }
+
+    @Inject(method = "tickAnimation", at = @At("HEAD"))
+    private void checkBlazeChef(CallbackInfo ci) {
         assert level != null;
         if (level.getBlockEntity(worldPosition.above()) instanceof HeatableBlockEntity)
-            return true;
-        if (aboveState.is(ModTags.HEAT_CONDUCTORS))
-            return level.getBlockEntity(worldPosition.above(2)) instanceof HeatableBlockEntity;
-        return false;
+            chef = true;
+        else if (level.getBlockState(worldPosition.above()).is(ModTags.HEAT_CONDUCTORS) &&
+                level.getBlockEntity(worldPosition.above(2)) instanceof HeatableBlockEntity)
+            chef = true;
+        else chef = false;
+    }
+
+    @ModifyExpressionValue(method = "tickAnimation", at = @At(value = "INVOKE", target = "Lcom/simibubi/create/content/processing/burner/BlazeBurnerBlockEntity;isValidBlockAbove()Z"))
+    private boolean isHeatableBlockEntityAbove(boolean original) {
+        return original || chef;
     }
 }